Add to the library the code to parse and generate scilab variable names.
The code is in an Util class that will need to be renamed someday but
it could as well stay like that for now.

Some basic unit tests are included. More are needed to speciffy exactly
what happens when non-convertible characters arise.
The convertion code is also used at multiple places within scilab but
variables names have additional limitations (they can't contain
mathematical signs for example) that need to be managed separatly from the
encoding.